Deduplicate code
