+++ cpmpy/tools/io/loader.py
@@ -19,7 +19,7 @@ from .dimacs import load_dimacs
 from cpmpy.tools.io.scip import load_scip
 from cpmpy.tools.io.wcnf import load_wcnf
 from cpmpy.tools.io.opb import load_opb
-from cpmpy.tools.io.utils import get_format
+from cpmpy.tools.io.utils import _derive_format
 
 # mapping format names to appropriate loader functions
 _loader_map = {
@@ -71,36 +71,7 @@ def load_formats() -> List[str]:
     """
     return list(_loader_map.keys())
 
-def _derive_format(file_path: str) -> str:
-    """
-    Derive the format of a file from its path by looking at its file extension.
 
-    Arguments:
-        file_path (str): The path to the file to derive the format from.
-
-    Raises:
-        ValueError: If the format could not be derived from the file path.
-
-    Returns:
-        The name of the format.
-
-    Example:
-        >>> _derive_format("instance.mps")
-        "mps"
-        >>> _derive_format("instance.lp.xz")
-        "lp"
-        >>> _derive_format("instance.cnf")
-        "dimacs"
-    """
-
-    # Iterate over the file path extensions in reverse order
-    for ext in file_path.split(".")[::-1]:
-        try:
-            return get_format(ext)
-        except ValueError:
-            continue
-
-    raise ValueError(f"No file format provided and could not derive format from file path: {file_path}")
 
 def load(file_path: str, format: Optional[str] = None) -> cp.Model:
     """
